add dataset halueval
